identify-unit/src/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
paho-mqtt
passlib[bcrypt]
bcrypt==3.1.7
jinja2
python-multipart